[fix] file path change
